slcd-backend-new/requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.7
dj-database-url==1.2.0
Django==3.2.15
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
drf-yasg==1.21.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
packaging==23.0
Pillow==9.2.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pytz==2022.2.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.15
whitenoise==6.3.0
